requirements.txt
# Kree AI — Python Dependencies
# Pin versions for reproducible builds.
# Run: pip install -r requirements.txt

# ── Core AI ───────────────────────────────────────────
google-genai>=1.0.0
google-generativeai>=0.8.0

# ── Audio & Media ────────────────────────────────────
pyaudio>=0.2.14
opencv-python>=4.9.0
pillow>=10.2.0
numpy>=1.26.0

# ── Desktop UI ───────────────────────────────────────
pywebview>=5.0

# ── System Control ───────────────────────────────────
pyautogui>=0.9.54
pyperclip>=1.8.2
mss>=9.0.0
psutil>=5.9.0
comtypes>=1.4.0
pycaw>=20240210

# ── Notifications & OS ───────────────────────────────
win10toast>=0.9
send2trash>=1.8.0

# ── Web & Scraping ───────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
youtube-transcript-api>=0.6.0

# ── Browser Automation ───────────────────────────────
playwright>=1.40.0

# ── QR Code (PWA Connect Tab) ────────────────────────
qrcode[pil]>=7.4.0

# ── Build Tools (dev only — not needed at runtime) ───
# pyinstaller>=6.0
# pyinstaller-hooks-contrib>=2024.0

# ── Production Missing Modules ──────────────────────
edge-tts>=6.1.12
openwakeword>=0.6.0
scipy>=1.12.0
pystray>=0.19.5
resemblyzer>=0.1.4
onnxruntime>=1.17.1
scikit-learn>=1.4.0